+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-07-17 21:04:36
-// Commit: feat(forms): add async forms stream provider
+// Updated at: 2026-07-20 17:23:51
+// Commit: feat(forms): integrate custom telemetry logger in forms
 
 class formsComponentState {
-  final String commitRef = '2923';
-  final String lastMessage = 'feat(forms): add async forms stream provider';
-  final DateTime timestamp = DateTime.parse('2026-07-17T21:04:36+05:00');
+  final String commitRef = '2953';
+  final String lastMessage = 'feat(forms): integrate custom telemetry logger in forms';
+  final DateTime timestamp = DateTime.parse('2026-07-20T17:23:51+05:00');
 }